警告つぶし。

+++ VCECore/rgy_output_avcodec.cpp
@@ -2283,7 +2283,7 @@ RGY_ERR RGYOutputAvcodec::WriteNextFrameInternal(RGYBitstream *pBitstream, int64
         const auto frameI = (pBitstream->frametype() & (RGY_FRAMETYPE_IDR | RGY_FRAMETYPE_I)) != 0;
         auto& qVideoQueueFree = (frameI) ? m_Mux.thread.qVideobitstreamFreeI : m_Mux.thread.qVideobitstreamFreePB;
         auto queueFavoredSize = (frameI) ? VID_BITSTREAM_QUEUE_SIZE_I : VID_BITSTREAM_QUEUE_SIZE_PB;
-        if (qVideoQueueFree.size() > queueFavoredSize) {
+        if (qVideoQueueFree.size() > (size_t)queueFavoredSize) {
             //あまり多すぎると無駄にメモリを使用するので減らす
             pBitstream->clear();
         } else {
